revisão textual

+++ disciplinas/aed2/programas/calculadora_generica_up/registro.c
@@ -6,13 +6,12 @@ Calculator* make_calculator(OPERATION_TYPE ot){
     switch (ot)
     {
     case INT:
-        return calculator_create(sizeof(int), 2, (const Operation[]){add_int, multiply_int});  
-    
+        return calculator_create(sizeof(int), 2, (const Operation[]){add_int, multiply_int});
+
     case DOUBLE:
-        return calculator_create(sizeof(int), 2, (const Operation[]){add_double, multiply_double});  
-    
+        return calculator_create(sizeof(double), 2, (const Operation[]){add_double, multiply_double});
+
     default:
         return NULL;
     }
 }
-
